pyproject.toml
[project]
name = "franq-orig"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "langchain>=1.2.10",
    "langchain-anthropic>=1.3.4",
    "langgraph>=1.0.9",
    "pandas>=2.0,<3",
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.47",
    "streamlit>=1.44.0",
]
